Add files via upload
This project is a phonebook web application built with JSP Servlets, utilizing a MySQL database for data storage. The front end is developed using Bootstrap, HTML, CSS, and Thymeleaf for templating. Users can manage their contacts through an intuitive and responsive interface, with data seamlessly stored and retrieved from the MySQL database. The integration of these technologies ensures a dynamic and user-friendly experience for managing and organizing contact information.